added initial divide method

+++ src/Calc.java
@@ -19,6 +19,6 @@ public class Calc
    }
    
    public static float divide(int a, int b) {
-	   return (float)a / b;
+	   return a / b;
    }
 }
